dev mode

+++ Bleach Rebirth of Souls Community Patch.py
@@ -49,9 +49,11 @@ try:
         except Exception:
             return "unknown"
     def pulling_from_git():
-        subprocess.run(["git","-C",BASE_DIR,"fetch"], check=True, capture_output=True, text=True)
-        subprocess.run(["git","-C",BASE_DIR,"reset","--hard","origin/main"], check=True, capture_output=True, text=True)
-        subprocess.run(["git","-C",BASE_DIR,"clean","-fd","-e","Json"], check=True, capture_output=True, text=True)
+        if os.path.exists(BASE_DIR,"BalanceLeadTools","DevToken.txt") == False:
+            subprocess.run(["git","-C",BASE_DIR,"fetch"], check=True, capture_output=True, text=True)
+            subprocess.run(["git","-C",BASE_DIR,"reset","--hard","origin/main"], check=True, capture_output=True, text=True)
+            subprocess.run(["git","-C",BASE_DIR,"clean","-fd","-e","Json"], check=True, capture_output=True, text=True)
+            print("not in dev mode")
         return subprocess.run(["git", "-C", BASE_DIR, "pull"], check=True, capture_output=True, text=True)
     try: 
         result = pulling_from_git()
